mysql 데이터베이스 연결, 게시글 페이징 처리

+++ backend/src/main/java/com/bbs/backend/controller/UserController.java
@@ -5,8 +5,10 @@ import com.bbs.backend.dto.user.LoginDTO;
 import com.bbs.backend.dto.user.UserDTO;
 import com.bbs.backend.dto.user.UserInfoDTO;
 import com.bbs.backend.entity.UserEntity;
+import com.bbs.backend.exception.UserNotFoundException;
 import com.bbs.backend.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +18,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.net.URI;
 import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,6 +29,8 @@ public class UserController {
     @PostMapping
     public ResponseEntity<?> createUser(@Validated @RequestBody UserDTO userDTO) {
         UserEntity userEntity = UserDTO.toEntity(userDTO);
+        userEntity.setId(UUID.randomUUID().toString());
+
         if (userService.saveUser(userEntity)) {
             URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                     .path("/yourAccount")
@@ -62,7 +67,11 @@ public class UserController {
 
     @GetMapping("/yourAccount")
     public ResponseEntity<UserInfoDTO> getUserInfo(@SessionAttribute(name = SessionConst.LOGIN_USER, required = false) String id) {
-        UserInfoDTO userInfoDTO = new UserInfoDTO(userService.getUserInfo(id));
+        UserEntity user = userService.getUserInfo(id);
+        if (user == null) {
+            throw new UserNotFoundException("User doesn't exist");
+        }
+        UserInfoDTO userInfoDTO = new UserInfoDTO(user);
 
         return ResponseEntity.ok(userInfoDTO);
     }
@@ -70,6 +79,10 @@ public class UserController {
     @DeleteMapping("/yourAccount")
     public ResponseEntity<?> deleteUser(HttpServletRequest request) {
         HttpSession session = request.getSession(false);
+        if (session == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         String id = (String)session.getAttribute(SessionConst.LOGIN_USER);
         userService.deleteUser(id);
         session.invalidate();
